Support uploading MSZIP files
Microsoft requires .CAB file to upload to Windows Hardware Dev Center, however
after the content is signed by Microsoft, it's unavailable in .CAB format from
Hardware Dev Center. It's instead distributed using .ZIP file.

Allow LVFS to accept the ZIP file format and repack it into CAB. This will make
it easier to re-use content otherwise headed to Windows Update.

Fixes https://github.com/hughsie/lvfs-website/issues/20
