+++ app/flaskapp.cfg
@@ -8,7 +8,6 @@ IP = '127.0.0.1'
 PORT = 8080
 DOWNLOAD_DIR = '/home/hughsie/Code/lvfs-website/downloads/'
 KEYRING_DIR = 'gnupg'
-CABEXTRACT_CMD = '/usr/bin/cabextract'
 CDN_URI = None
 CDN_BUCKET = None
 DATABASE_HOST = 'localhost'
